challenge/requirements.txt
torch
transformers
datasets
faiss-cpu
sentence-transformers
tqdm
Pillow
matplotlib
scikit-learn
gdown
